👌 [RUMF-1267] collocate remaining types from domain/record/types
