+++ packages/rum/src/domain/record/index.ts
@@ -1,3 +1,2 @@
 export { record } from './record'
 export { serializeNodeWithId, serializeDocument } from './serialize'
-export * from './types'
