+++ packages/rum/src/domain/record/mutationObserver.ts
@@ -12,17 +12,37 @@ import {
   nodeAndAncestorsHaveSerializedNode,
 } from './serializationUtils'
 import { serializeNodeWithId, serializeAttribute } from './serialize'
-import type {
-  RumAttributesMutationRecord,
-  RumCharacterDataMutationRecord,
-  RumChildListMutationRecord,
-  MutationCallBack,
-  RumMutationRecord,
-} from './types'
 import { forEach } from './utils'
 import { createMutationBatch } from './mutationBatch'
+import type { MutationCallBack } from './observer'
 
 type WithSerializedTarget<T> = T & { target: NodeWithSerializedNode }
+
+// https://dom.spec.whatwg.org/#interface-mutationrecord
+interface RumCharacterDataMutationRecord {
+  type: 'characterData'
+  target: Node
+  oldValue: string | null
+}
+
+interface RumAttributesMutationRecord {
+  type: 'attributes'
+  target: Element
+  oldValue: string | null
+  attributeName: string | null
+}
+
+interface RumChildListMutationRecord {
+  type: 'childList'
+  target: Node
+  addedNodes: NodeList
+  removedNodes: NodeList
+}
+
+export type RumMutationRecord =
+  | RumCharacterDataMutationRecord
+  | RumAttributesMutationRecord
+  | RumChildListMutationRecord
 
 /**
  * Buffers and aggregate mutations generated by a MutationObserver into MutationPayload
